chore(mitto-ayl): document the 30s glob cache TTL and tidy the walk counters (refs mitto-ayl)
Review-phase fix-ups for mitto-ayl.

.augment/rules/07-prompts.md still documented the pre-mitto-ayl "memoised for
5 s per (folder, pattern, wantFiles)" figure for glob-mode
FileExists/DirExists. Updated to 30 s and extended to mention that concurrent
misses on the same key now collapse into a single walk via singleflight, with
the reason the old TTL never hit at the observed workspace-prompts request
rate. docs/devel/prompt-templates.md and docs/config/prompts.md were checked
and never carried the glob TTL figure, so they need no change.

In the two new singleflight tests, walkCount was declared int32 while being
guarded by a sync.Mutex; that falsely implied sync/atomic access. Narrowed to
a plain int since the mutex already provides the synchronisation.

No production behaviour change. go build ./..., make lint-go (0 issues),
gofmt, go vet, and go test ./internal/cel/... -race plus
./internal/prompts/... ./internal/config/... ./internal/pathglob/... all pass.

+++ internal/cel/templatefuncs_glob_test.go
@@ -444,7 +444,7 @@ func TestGlob_Singleflight_CollapsesConcurrentWalks(t *testing.T) {
 	root := makeGlobFixture(t)
 
 	origWalk := walkMatchFn
-	var walkCount int32
+	var walkCount int
 	var mu sync.Mutex
 	walkMatchFn = func(opts pathglob.WalkMatchOpts) pathglob.WalkMatchResult {
 		mu.Lock()
@@ -485,7 +485,7 @@ func TestGlob_Singleflight_DistinctKeysWalkIndependently(t *testing.T) {
 	root := makeGlobFixture(t)
 
 	origWalk := walkMatchFn
-	var walkCount int32
+	var walkCount int
 	var mu sync.Mutex
 	walkMatchFn = func(opts pathglob.WalkMatchOpts) pathglob.WalkMatchResult {
 		mu.Lock()
@@ -511,7 +511,7 @@ func TestGlob_Singleflight_DistinctKeysWalkIndependently(t *testing.T) {
 	mu.Lock()
 	got := walkCount
 	mu.Unlock()
-	if got != int32(len(patterns)) {
+	if got != len(patterns) {
 		t.Errorf("WalkMatch call count for %d distinct keys = %d, want %d (distinct keys must not collapse)", len(patterns), got, len(patterns))
 	}
 }
